% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_cols.R
\name{my_cols}
\alias{my_cols}
\title{Colors from famous academic journals}
\usage{
my_cols(...)
}
\arguments{
\item{...}{the name of color, such as "jama_1", "nejm_2".}
}
\description{
This function provides useful color sets of famous academic journals, including NEJM, JAMA, and Nature.
}
\examples{
require(ggplot2)
data(diamonds)
set.seed(42)
small <- diamonds[sample(nrow(diamonds), 1000), ]
ggplot(data = small) +
 geom_point(mapping = aes(x = carat, y = price), col=my_cols("jama_2"))
}
