% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_cn.R
\name{set_cn}
\alias{set_cn}
\title{Set Chinese Support Environment for Rstudio}
\usage{
set_cn()
}
\arguments{
\item{...}{There is no parameters}
}
\value{

}
\description{
This function set Chinese support environment for Rstudio
}
